Refactor hash calculation to be a function on HashAlgorithm
It seems more natural to bundle hash calculation on different entities,
like maybe strings in the future, in this class.
